have the weather api displaying + fixed merge conflicts

+++ WeekendWarrior/Trail.swift
@@ -13,8 +13,14 @@ struct Trail {
     var description: String? {return json["description"] as? String}
     var id: String? {return json["objectID"] as? String}
     
-    var lat: Float? {return json["lat"] as? Float}
-    var lng: Float? {return json["lng"] as? Float}
+    //var _geoloc: String? {return json["_geoloc"] as? String}
+    
+    var lat: Double? {
+        // let geoloc = json["_geoloc"] as? [String:Any]
+        // return geoloc!["lat"] as! Double
+        return json["lat"] as? Double
+    }
+    var lng: Double? {return json["lng"] as? Double}
     
 
 }
